Rename fetch list methods
